+++ app/src/main/java/com/rsupport/mobile1/test/activity/data/model/GettyImage.kt
@@ -1,4 +1,4 @@
-package com.rsupport.mobile1.test.activity
+package com.rsupport.mobile1.test.activity.data.model
 
 data class GettyImage(
     val id: String,
