+++ app/src/main/java/com/rsupport/mobile1/test/activity/main/MainActivity.kt
@@ -1,9 +1,10 @@
-package com.rsupport.mobile1.test.activity
+package com.rsupport.mobile1.test.activity.main
 
 import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.commit
+import com.rsupport.mobile1.test.activity.GettyImageViewModel
 import com.rsupport.mobile1.test.databinding.ActivityMainBinding
 import dagger.hilt.android.AndroidEntryPoint
