+++ app/src/main/java/com/rsupport/mobile1/test/activity/util/ViewUtil.kt
@@ -1,4 +1,4 @@
-package com.rsupport.mobile1.test.activity
+package com.rsupport.mobile1.test.activity.util
 
 import android.content.res.Resources
 import android.util.DisplayMetrics
