+++ app/src/main/res/layout/fragment_getty_image.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.GettyImageFragment"
+    tools:context=".activity.main.GettyImageFragment"
     tools:ignore="SpUsage">
 
     <androidx.recyclerview.widget.RecyclerView
